Fixed a method.

+++ messaging/GLPUserDao.h
@@ -15,7 +15,7 @@
 + (GLPUser *)findByKey:(NSInteger)key db:(FMDatabase *)db;
 + (GLPUser *)findByRemoteKey:(NSInteger)remoteKey db:(FMDatabase *)db;
 + (void)save:(GLPUser *)entity inDb:(FMDatabase *)db;
-+ (int)saveIfNotExist:(GLPUser*)entity db:(FMDatabase *)db;
++ (NSInteger)saveIfNotExist:(GLPUser*)entity db:(FMDatabase *)db;
 + (void)saveUsersIfNotExists:(NSArray *)users db:(FMDatabase *)db;
 + (void)updateLoggedInUsersName:(NSString *)name andSurname:(NSString *)surname;
 + (void)updateLoggedInUsersTagline:(NSString *)tagline;
